pyproject.toml
[project]
name = "vdmc-salesforce-migration"
version = "0.1.0"
description = "Internal library for Salesforce data migration"
authors = [
    { name="Pascal Klos", email="pascal.klos@dmc.valantic.com" }
]
readme = "README.md"
requires-python = ">=3.10"

dependencies = [
    "simple-salesforce>=1.12.5",
    "pandas>=1.5",
    "requests>=2.28",
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "black",
    "ruff",
]

[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"